Fixed conversion warning.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@93881 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/objtools/data_loaders/genbank/impl/psg_loader_impl.hpp
@@ -116,7 +116,7 @@ struct SPsgBlobInfo
     int blob_state;
     Int8 last_modified;
 
-    int GetBlobVersion() const { return last_modified/60000; /* ms to minutes */ }
+    int GetBlobVersion() const { return int(last_modified/60000); /* ms to minutes */ }
 
     bool IsSplit() const { return !id2_info.empty(); }
